cleaned

+++ src/objects/plane/plane.c
@@ -25,7 +25,7 @@ t_plane	*plane(t_fvector3 position, t_fvector3 normal, t_rgb color)
 	pl->position = position;
 	pl->normal = ft_fnormalize(normal);
 	ref = (t_fvector3){0, 1, 0};
-	if (fabs(ft_fdot_product(pl->normal, ref)) > 0.999f)
+	if (fabsf(ft_fdot_product(pl->normal, ref)) > 0.999f)
 		ref = (t_fvector3){1, 0, 0};
 	pl->right = ft_fnormalize(ft_fcross_product(pl->normal, ref));
 	pl->up = ft_fcross_product(pl->right, pl->normal);
